+++ app/javascript/users/components/homepage/sections/Affiliates.vue
@@ -8,28 +8,31 @@
       <base-section-heading title="Chi sta usando Insiemento?" />
 
       <v-slide-group>
-        <template v-for="n in 2">
+        <template v-for="(affiliate, index) in affiliates">
           <v-slide-item
-            :key="n"
+            :key="index"
             class="align-self-center"
           >
-            <base-img
-              :src="require(`../../../assets/images/logo-affiliate-${n}.png`)"
-              color="grey"
-              contain
-              height="40"
-              width="128"
-            />
+            <a :href="affiliate.url">
+              <base-img
+                  :src="require(`../../../assets/images/logo-affiliate-${affiliate.name}.png`)"
+                  color="grey"
+                  contain
+                  :height="affiliate.height"
+                  :width="affiliate.width"
+              />
+            </a>
+
           </v-slide-item>
 
           <v-responsive
-            v-if="n < 6"
-            :key="`divider-${n}`"
+            v-if="index < 6"
+            :key="`divider-${index}`"
             class="text-center"
             height="56"
             width="48"
           >
-            <v-divider vertical />
+            <v-divider vertical v-if="index+1 < affiliates.length"/>
           </v-responsive>
         </template>
       </v-slide-group>
@@ -40,5 +43,14 @@
 <script>
   export default {
     name: 'SectionAffiliates',
+    computed: {
+      affiliates() {
+        return [
+          { name: 'fitness' , height: '50', width: '120', url: null },
+          { name: 'dsm'     , height: '100', width: '100', url: 'https://www.dsmdanceschool.it' },
+          { name: 'victor'  , height: '70', width: '70', url: null },
+        ]
+      },
+    }
   }
 </script>